+++ lib/src/token/src/token.dart
@@ -36,7 +36,7 @@ abstract class Token with _$Token {
     /// [onResponse] is the callback to call when the response is received
     void Function(String statusCode)? onResponse,
   }) async {
-    final connector = LayrzConnector(uri: uri);
+    final connector = LayrzConnector(uri: uri, apiToken: apiToken);
     try {
       final response = await connector.perform(
         GqlQuery(
@@ -94,7 +94,7 @@ abstract class Token with _$Token {
     /// [onResponse] is the callback to call when the response is received
     void Function(String statusCode)? onResponse,
   }) async {
-    final connector = LayrzConnector(uri: uri);
+    final connector = LayrzConnector(uri: uri, apiToken: apiToken);
 
     try {
       final response = await connector.perform(
@@ -155,7 +155,7 @@ abstract class Token with _$Token {
     /// [onResponse] is the callback to call when the response is received
     void Function(String statusCode)? onResponse,
   }) async {
-    final connector = LayrzConnector(uri: uri);
+    final connector = LayrzConnector(uri: uri, apiToken: apiToken);
 
     try {
       final response = await connector.perform(